refactor: use u64 instead of int for timeouts

+++ include/solana.hpp
@@ -514,7 +514,7 @@ class Connection {
   getSignatureStatuses(const std::vector<std::string> &signatures,
                        bool searchTransactionHistory = false) const;
   
-  bool confirmTransaction(std::string transactionSignature,int timeout, std::string confirmLevel) const;
+  bool confirmTransaction(std::string transactionSignature,uint64_t timeout, std::string confirmLevel) const;
 
   /**
    * Fetch the current status of a signature
